requirements.txt
# Web Framework
streamlit>=1.28.0
streamlit-webrtc>=0.47.0

# Video Processing
ffmpeg-python>=0.2.0
opencv-python-headless>=4.8.0

# Audio/ASR
openai-whisper>=20231117

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# LLM API
openai>=1.3.0
anthropic>=0.18.0
requests>=2.31.0

# Config & Security
cryptography>=41.0.0
python-dotenv>=1.0.0

# Utilities
numpy>=1.24.0
scikit-image>=0.21.0
tqdm>=4.66.0
